WIP: EXPERIMENT to time the improvement of vectorcall
This tags on vectorcall support the terrible way by inserting things
manually after cython is done.
Very brittle, but allows timing that this saves a huge amount of
overhead, making the call as fast WITH keyword arguments as it is
normally WITHOUT them (and also boosting it without).

That can save a whopping 200ns or ~40% of the total call overhead