MAINT: Simplify caching logic to a single cache and some cleanups
There is no reason to have a priority-cache that gets invalidated,
we can just store the final `(backend, impl)` we found and the
`_max_priority` when it occurred in the normal type-cache directly.

Saves a dict-lookup, and the hassle of two caches.  It was really
rather silly how it was to start with :).

Also does some cleanup, adds a bit of docs, and returns the `None`
backend if there is a fallback.